requirements.txt
absl-py==1.1.0
aiobotocore==2.3.4
aiohttp==3.8.1
aioitertools==0.10.0
aiosignal==1.2.0
alabaster==0.7.12
alembic==1.7.5
altair==4.2.0
anyio==3.5.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.6.6
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.5.7
azure-core==1.25.0
azure-storage-blob==12.5.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
black==22.3.0
bleach==4.1.0
blinker==1.4
bokeh==2.3.3
boto3==1.21.21
botocore==1.24.21
Bottleneck==1.3.5
branca==0.5.0
bravado==11.0.3
bravado-core==5.17.0
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.2.0
certifi==2022.6.15
certipy==0.1.3
cffi==1.14.6
charset-normalizer==2.0.4
click==8.1.3
cloudpickle==1.6.0
colorama==0.4.5
commonmark==0.9.1
conda==4.9.2
conda-package-handling==1.8.1
coverage==6.4.3
coveralls==3.3.1
cryptography==37.0.1
cycler==0.11.0
Cython==0.29.30
cytoolz==0.11.0
dask==2021.3.1
datascience==0.17.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
distlib==0.3.5
distributed==2021.4.1
docopt==0.6.2
docutils==0.19
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.1
filelock==3.7.1
flake8==3.9.2
Flask==2.2.2
flatbuffers==2.0
flatten-dict==0.4.2
folium==0.12.1.post1
frozenlist==1.3.1
fsspec==2022.3.0
future==0.18.2
gast==0.4.0
gevent==21.12.0
gitdb==4.0.9
GitPython==3.1.27
gmpy2==2.1.2
gofer-grader==1.1.0
google-auth==2.9.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.11
greenlet==1.1.1
grpcio==1.47.0
h5py==3.1.0
HeapDict==1.0.1
httpie==2.5.0
idna==3.3
imageio==2.19.3
imagesize==1.4.1
importlib-metadata==4.11.3
importlib-resources==5.2.0
iniconfig==1.1.1
ipykernel==6.14.0
ipyleaflet==0.17.0
ipympl==0.6.3
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isodate==0.6.1
isort==5.10.1
itsdangerous==2.1.2
jedi==0.17.2
Jinja2==3.0.3
jmespath==1.0.1
joblib==1.1.0
json5==0.9.6
jsonpickle==1.5.2
jsonpointer==2.3
jsonref==0.2
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.10.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-lsp==1.4.1
jupyter-nbextensions-configurator==0.4.1
jupyter-server==1.18.1
jupyter-server-mathjax==0.2.5
jupyter-server-proxy==3.1.0
jupyter-telemetry==0.1.0
jupyter-tensorboard==0.2.0
jupyterhub==1.3.0
jupyterlab==3.1.14
jupyterlab-code-formatter==1.4.10
jupyterlab-drawio==0.9.0
jupyterlab-git==0.32.4
jupyterlab-lsp==3.9.1
jupyterlab-materialdarker==0.6.0
jupyterlab-pygments==0.1.2
jupyterlab-s3-browser==0.10.1
jupyterlab-server==2.12.0
jupyterlab-topbar==0.6.1
jupyterlab-widgets==1.0.2
keras==2.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
lazy-object-proxy==1.7.1
libclang==14.0.1
lime==0.2.0.1
llvmlite==0.36.0
locket==1.0.0
lxml==4.9.1
Mako==1.1.4
mamba==0.1.2
mamba-gator==5.2.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.3.4
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
mock==4.0.3
monotonic==1.6
mpmath==1.2.1
msgpack==1.0.3
msrest==0.7.1
multidict==6.0.2
munch==2.5.0
mypy-extensions==0.4.3
nb-conda-kernels==2.3.1
nbclassic==0.3.5
nbclient==0.5.13
nbconvert==6.4.4
nbdime==3.1.1
nbformat==5.3.0
nbgitpuller==0.9.0
nbpresent==3.0.0
nbresuse==0.4.0
nbsphinx==0.8.9
nbstripout==0.6.0
neptune-client==0.16.4
neptune-sacred==0.9.7
neptune-sklearn==0.9.6
neptune-tensorboard==0.5.1
nest-asyncio==1.5.5
networkx==2.8.4
nltk==3.5
notebook==6.2.0
numba==0.53.1
numexpr==2.7.3
numpy==1.22.1
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pamela==1.0.0
pandas==1.2.5
pandocfilters==1.5.0
parso==0.7.1
partd==1.2.0
pathspec==0.9.0
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.2.2
pipenv==2022.8.5
platformdirs==2.5.2
plotly==5.3.1
pluggy==0.13.1
pooch==1.4.0
prometheus-client==0.14.1
prompt-toolkit==3.0.20
protobuf==3.15.8
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-cpuinfo==8.0.0
py4j==0.10.7
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.2
pybind11-global==2.9.2
pycodestyle==2.7.0
pycosat==0.6.3
pycparser==2.21
pycurl==7.43.0.6
pydeck==0.7.1
pydocstyle==6.1.1
pyflakes==2.3.1
Pygments==2.11.2
PyJWT==2.4.0
pylint==2.9.6
pymongo==4.2.0
Pympler==1.0.1
pyodbc==4.0.34
pyOpenSSL==22.0.0
pyparsing==3.0.4
pyrsistent==0.18.0
PySocks==1.7.1
pytest==6.1.1
pytest-cov==2.10.1
python-dateutil==2.8.2
python-json-logger==2.0.1
python-jsonrpc-server==0.4.0
python-language-server==0.26.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.2
pytoolconfig==1.2.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.0
regex==2022.7.25
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3987==1.3.8
rich==12.5.1
rise==5.7.1
rope==1.3.0
rpy2==3.4.5
rsa==4.8
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.6
ruamel-yaml-conda==0.15.100
s3fs==0.5.1
s3transfer==0.5.2
sacred==0.8.2
sacredboard==0.4.2
scikit-image==0.18.3
scikit-learn==1.1.1
scikit-plot==0.3.7
scipy==1.6.3
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
setuptools==59.8.0
shap==0.40.0
simpervisor==0.4
simplegeneric==0.8.1
simplejson==3.17.6
singleton-decorator==1.0.0
six==1.16.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
Sphinx==5.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.39
stack-data==0.2.0
statsmodels==0.12.2
streamlit==1.11.1
strict-rfc3339==0.7
swagger-spec-validator==2.7.4
sympy==1.7.1
tables==3.6.1
tblib==1.7.0
tenacity==8.0.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.13.1
testpath==0.6.0
threadpoolctl==2.2.0
tifffile==2020.10.1
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
torch==1.10.0+cu113
torchvision==0.11.1+cu113
torchviz==0.0.2
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
traittypes==0.2.1
typing_extensions==4.3.0
tzlocal==2.1
ujson==5.4.0
urllib3==1.26.11
validators==0.20.0
vincent==0.4.4
virtualenv==20.16.3
virtualenv-clone==0.5.7
voila==0.2.15
watchdog==2.1.9
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.2.2
wheel==0.37.1
widgetsnbextension==3.5.2
wrapt==1.12.1
xlrd==2.0.1
xyzservices==2022.6.0
yapf==0.31.0
yarl==1.8.1
yellowbrick==1.4
zict==2.1.0
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
